+++ src/main/kotlin/com/zepben/evolve/database/postgres/metrics/MetricsDatabaseTables.kt
@@ -12,7 +12,7 @@ import com.zepben.evolve.database.postgres.metrics.tables.TableJobSources
 import com.zepben.evolve.database.postgres.metrics.tables.TableJobs
 import com.zepben.evolve.database.postgres.metrics.tables.TableNetworkContainerMetrics
 import com.zepben.evolve.database.postgres.metrics.tables.tableMetricsVersion
-import com.zepben.evolve.database.sql.tables.SqlTable
+import com.zepben.evolve.database.sql.SqlTable
 import com.zepben.evolve.database.sql.BaseDatabaseTables
 
 /**