+++ src/main/kotlin/com/zepben/evolve/database/postgres/metrics/tables/TableJobSources.kt
@@ -8,9 +8,9 @@
 
 package com.zepben.evolve.database.postgres.metrics.tables
 
-import com.zepben.evolve.database.sql.tables.Column
-import com.zepben.evolve.database.sql.tables.Column.Nullable.NOT_NULL
-import com.zepben.evolve.database.sql.tables.Column.Nullable.NULL
+import com.zepben.evolve.database.sql.Column
+import com.zepben.evolve.database.sql.Column.Nullable.NOT_NULL
+import com.zepben.evolve.database.sql.Column.Nullable.NULL
 
 @Suppress("PropertyName")
 class TableJobSources : MultiJobTable() {