+++ src/main/kotlin/com/zepben/evolve/database/sqlite/cim/tables/iec61970/base/wires/TableEnergySourcePhases.kt
@@ -8,8 +8,8 @@
 
 package com.zepben.evolve.database.sqlite.cim.tables.iec61970.base.wires
 
-import com.zepben.evolve.database.sql.tables.Column
-import com.zepben.evolve.database.sql.tables.Column.Nullable.NOT_NULL
+import com.zepben.evolve.database.sql.Column
+import com.zepben.evolve.database.sql.Column.Nullable.NOT_NULL
 import com.zepben.evolve.database.sqlite.cim.tables.iec61970.base.core.TablePowerSystemResources
 
 @Suppress("PropertyName")